Don't show the "EXA/XAA" switch by default on Intel cards
Newer intel drivers properly autoselect UXA/EXA if KMS is available or
not, and XAA should really only be used as a fallback. For now, don't
set EXA => 0 in the default values for the Intel cards. This will hide
the XAA/EXA switch from the configuration UI. If an accel method was
explicitly selected in xorg.conf by the user, the UI will be shown so
that he has a chance to unselect EXA.

git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/drakx-kbd-mouse-x11/trunk@256914 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94